Finished first exam prep with new MVCServer

+++ CSharp-Web-Server/Apps/Git/Controllers/CommitsController.cs
@@ -52,6 +52,10 @@
         public HttpResponse Delete(string id)
         {
             var userId = this.User.Id;
+            if (userId != id)
+            {
+                return this.Error("You dont have access to remove this repository!");
+            }
             this.commitsService.RemoveCommit(userId);
             return this.Redirect("/Commits/All");
         }
